Add support for fractional FPS.

+++ esp32/program.hpp
@@ -14,7 +14,7 @@ extern uint16_t autonomous_speed;
 class Program
 {
 public:
-    Program(int scale)
+    Program(float scale)
         : limiter(scale)
     {
     }
